+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Dev
+
+* Add `parallelLimit` option that allows to limit number of browsers
+ run in parallel.
+
 ## 0.4.1 - 2014-06-19
 
 * Correctly detect crop region in Firefox